requirements.txt
fastmcp>=2.0.0
pydantic>=2.0.0
aiomysql>=0.2.0
python-dotenv>=1.0.0
